requirements.txt
google-genai
requests
beautifulsoup4
markdown
google-auth
google-auth-oauthlib
google-api-python-client
moviepy==1.0.3
numpy<2.0.0
pillow
imageio-ffmpeg
decorator
tqdm
PyGithub
feedparser
selenium
webdriver-manager
trafilatura
tenacity
json_repair
regex
pydantic
opencv-python-headless
numpy
openai
oauth2client
puter
